requirements.txt
requests==2.25.1
google.cloud